Refactor cpu_usage code out of spammer.go
